+++ src/REPLMode/command_declarations.jl
@@ -51,17 +51,21 @@ compound_declarations = [
                 PSA[:name => "verbose", :short_name => "v", :api => :verbose => true],
                 PSA[:name => "workspace", :api => :workspace => true],
                 PSA[:name => "julia_version_strict", :api => :julia_version_strict => false],
+                PSA[:name => "update_on_mismatch", :short_name => "u", :api => :update_on_mismatch => true],
             ],
             :description => "downloads all the dependencies for the project",
             :help => md"""
-                    instantiate [-v|--verbose] [--workspace] [--julia_version_strict]
-                    instantiate [-v|--verbose] [--workspace] [--julia_version_strict] [-m|--manifest]
-                    instantiate [-v|--verbose] [--workspace] [--julia_version_strict] [-p|--project]
+                    instantiate [-v|--verbose] [--workspace] [--julia_version_strict] [-u|--update_on_mismatch]
+                    instantiate [-v|--verbose] [--workspace] [--julia_version_strict] [-u|--update_on_mismatch] [-m|--manifest]
+                    instantiate [-v|--verbose] [--workspace] [--julia_version_strict] [-u|--update_on_mismatch] [-p|--project]
 
                 Download all the dependencies for the current project at the version given by the project's manifest.
                 If no manifest exists or the `--project` option is given, resolve and download the dependencies compatible with the project.
                 If `--workspace` is given, all dependencies in the workspace will be downloaded.
                 If `--julia_version_strict` is given, manifest version check failures will error instead of log warnings.
+                If `-u`/`--update_on_mismatch` is given, fall back to `pkg> update` when the manifest does not match the project or
+                was resolved with a different Julia minor version, instead of warning or erroring. Useful for tooling environments
+                where any compatible set of dependency versions is acceptable.
 
                 After packages have been installed the project will be precompiled. For more information see `pkg> ?precompile`.
                 """,